Migrate u/v -> gamma/beta tests
