+++ src/python/zquantum/qaoa/parameter_initialization/__init__.py
@@ -1,3 +1,3 @@
-from ._fourier import FourierOptimizer
+from ._fourier import FourierOptimizer, convert_u_v_to_gamma_beta
 from ._interp import get_new_layer_params_using_interp
 from ._old_fourier import Fourier
